Simplified `stringify' at a small convenience cost.
Previously `stringify' didn't behave exactly as intuitively expected,
largely in order to handle the idea that sometimes a number appearing
in a tree structure passed to it is really a character of a
string... and sometimes it's a number. For instance, hopefuly
`stringify' behavior would have been

stringify(["foo", 10, "bar"]) ---> "foo 10 bar"

but that would probably also indicate that

stringify([60, 61, 62]) ---> "<=>"

since it's not easy to tell if that's supposed to be a string or a
list of numbers. While workarounds are possible, the inclusion of
units is good practice anyway so now numbers are forbidden!

Also included is an extensive set of unit tests.